Completed first iteration of graph. Building tools are still not completely done.
The graph needs to have a color scale for the fpkm standard deviation, where 0-1 is good (green), and high is bad (red). The high end can be set with a spinner, possibly low as well. Also add a slider for the gap inbetween samples.

The standard deviation needs to be pulled out of the datafile and added to our space separated list, then pulled out and stored in each node. A color function needs to be written as well.